saga: move take first function to utils folder
