+++ app/utils/sagas/take-first.js
@@ -1,9 +1,10 @@
 import { fork, call, take } from 'redux-saga/effects';
 
-export default (pattern, saga, ...args) =>
-  fork(function*() {
+export default function takeFirst(pattern, saga, ...args) {
+  return fork(function*() {
     while (true) {
       const action = yield take(pattern);
       yield call(saga, ...args.concat(action));
     }
   });
+}
